# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Product Inventory API
  description: |
    A comprehensive hybrid ORM API for product inventory management using ASP.NET Core, 
    Entity Framework Core, and SQL Server stored procedures.
    
    ## Features
    - **Hybrid ORM Architecture**: Combines Entity Framework Core with stored procedures
    - **JWT Authentication**: Secure token-based authentication
    - **Role-based Authorization**: Fine-grained access control
    - **Analytics & Reporting**: Advanced analytics using stored procedures
    - **Inventory Management**: Complete inventory tracking and movement recording
    - **Health Checks**: Built-in monitoring and health endpoints
    
    ## Architecture
    The API implements Clean Architecture principles with:
    - **Entity Framework Core**: Standard CRUD operations and relationships
    - **Stored Procedures**: Complex analytics and reporting
    - **Raw SQL**: Performance-critical operations
    - **Repository Pattern**: Abstracted data access layer
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@productinventory.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.productinventory.com/api/v1
    description: Production server
  - url: https://staging-api.productinventory.com/api/v1
    description: Staging server
  - url: https://localhost:5001/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a specific product by its ID with inventory details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'

    put:
      tags:
        - Products
      summary: Update product
      description: Update an existing product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'

    delete:
      tags:
        - Products
      summary: Delete product
      description: Delete a product (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Product deleted successfully

  # Analytics Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the login endpoint
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Optional success/error message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        correlationId:
          type: string
          format: uuid
          description: Request correlation ID for tracking
      required:
        - success
        - timestamp
        - correlationId
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Category ID
        name:
          type: string
          description: Category name
        description:
          type: string
          description: Category description
        isActive:
          type: boolean
          description: Whether the category is active
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - name
        - isActive
        - createdAt
        - updatedAt
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Product ID
        name:
          type: string
          description: Product name
        description:
          type: string
          description: Product description
        sku:
          type: string
          description: Stock Keeping Unit
        price:
          type: number
          format: decimal
          description: Product price
        categoryId:
          type: integer
          format: int32
          description: Category ID
        category:
          $ref: '#/components/schemas/Category'
        inventory:
          $ref: '#/components/schemas/Inventory'
        isActive:
          type: boolean
          description: Whether the product is active
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - name
        - sku
        - price
        - categoryId
        - isActive
        - createdAt
        - updatedAt
...
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Product name
        description:
          type: string
          maxLength: 1000
          description: Product description
        sku:
          type: string
          minLength: 1
          maxLength: 50
          description: Stock Keeping Unit
        price:
          type: number
          format: decimal
          minimum: 0
          description: Product price
        categoryId:
          type: integer
          format: int32
          description: Category ID
        inventory:
          $ref: '#/components/schemas/UpdateInventoryRequest'
        isActive:
          type: boolean
          description: Whether the product is active

    # Inventory Schemas
    Inventory:
      type: object
      properties:
        productId:
          type: integer
          format: int32
          description: Product ID
        currentStock:
          type: integer
          format: int32
          description: Current stock level
        reorderPoint:
          type: integer
          format: int32
          description: Reorder point threshold
        maxStockLevel:
          type: integer
          format: int32
          description: Maximum stock level
        isInStock:
          type: boolean
          description: Whether the product is in stock
        lastUpdated:
          type: string
          format: date-time
          description: Last inventory update
      required:
        - productId
        - currentStock
        - reorderPoint
        - maxStockLevel
        - isInStock
        - lastUpdated
...
    UpdateInventoryRequest:
      type: object
      properties:
        currentStock:
          type: integer
          format: int32
          minimum: 0
          description: Current stock level
        reorderPoint:
          type: integer
          format: int32
          minimum: 0
          description: Reorder point threshold
        maxStockLevel:
          type: integer
          format: int32
          minimum: 0
          description: Maximum stock level

    # Inventory Movement Schemas
...